+++ app/src/main/AndroidManifest.xml
@@ -14,19 +14,19 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ONGAndroid">
         <activity
-            android:name=".view.SplashActivity"
+            android:name=".view.splash.SplashActivity"
             android:exported="false" />
         <activity
-            android:name=".view.LoginActivity"
-            android:exported="false" />
-        <activity
-            android:name=".view.MainActivity"
+            android:name=".view.login.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".view.home.MainActivity"
+            android:exported="false"/>
     </application>
 
 </manifest>
