--- app/src/main/java/com/melvin/ongandroid/view/LoginActivity.kt
@@ -1,12 +1,0 @@
-package com.melvin.ongandroid.view
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.melvin.ongandroid.R
-
-class LoginActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-    }
-}
