--- app/src/main/java/com/melvin/ongandroid/view/MainActivity.kt
@@ -1,30 +1,0 @@
-package com.melvin.ongandroid.view
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.NavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import com.melvin.ongandroid.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-    private lateinit var navController: NavController
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        // Retrieve NavController from the NavHostFragment
-        val navHostFragment = binding.navHostFragment.getFragment() as NavHostFragment
-        navController = navHostFragment.navController
-
-        //  binding.bnvMainNavigation.setOnNavigationItemSelectedListener(this)
-        binding.bnvMainNavigation.setupWithNavController(navController)
-    }
-}
