--- app/src/main/java/com/melvin/ongandroid/view/fragments/loginFragment/LoginFragment.kt
@@ -1,23 +1,0 @@
-package com.melvin.ongandroid.view.fragments.loginFragment
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.melvin.ongandroid.databinding.FragmentLoginBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class LoginFragment : Fragment() {
-    private var _binding: FragmentLoginBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View {
-        _binding = FragmentLoginBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-}
