+++ app/src/main/java/com/melvin/ongandroid/view/login/LoginActivity.kt
@@ -1,0 +1,19 @@
+package com.melvin.ongandroid.view.login
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.melvin.ongandroid.databinding.ActivityLoginBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LoginActivity : AppCompatActivity() {
+
+    private var _binding: ActivityLoginBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        _binding = ActivityLoginBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
