+++ app/src/main/java/com/melvin/ongandroid/view/login/fragments/LoginFragment.kt
@@ -1,0 +1,34 @@
+package com.melvin.ongandroid.view.login.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.Navigation
+import com.melvin.ongandroid.R
+import com.melvin.ongandroid.databinding.FragmentLoginBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LoginFragment : Fragment() {
+    private var _binding: FragmentLoginBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        _binding = FragmentLoginBinding.inflate(inflater, container, false)
+
+        onClickSignUp()
+
+        return binding.root
+    }
+
+    private fun onClickSignUp() {
+        binding.loginRegisterTxt.setOnClickListener {
+            Navigation.findNavController(it).navigate(R.id.action_loginFragment_to_signUpFragment)
+        }
+    }
+}
