+++ app/src/main/java/com/melvin/ongandroid/view/splash/SplashActivity.kt
@@ -1,0 +1,12 @@
+package com.melvin.ongandroid.view.splash
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.melvin.ongandroid.R
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+    }
+}
