+++ app/src/main/res/layout/fragment_activities.xml
@@ -9,7 +9,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.fragments.activitiesFragment.ActivitiesFragment">
+        tools:context=".view.home.fragments.activitiesFragment.ActivitiesFragment">
 
         <include
             android:id="@+id/iv_logo"
