+++ app/src/main/res/layout/fragment_contact.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.fragments.contactFragment.ContactFragment">
+        tools:context=".view.home.fragments.contactFragment.ContactFragment">
 
         <include
             android:id="@+id/ivLogo"
