+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.fragments.homeFragment.HomeFragment">
+    tools:context=".view.home.fragments.homeFragment.HomeFragment">
 
     <data>
         <!-- dataBinding (delete this after using this space) -->
