+++ app/src/main/res/layout/fragment_news.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.fragments.newsFragment.NewsFragment">
+    tools:context=".view.home.fragments.newsFragment.NewsFragment">
 
     <data></data>
