+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.melvin.ongandroid.view.fragments.homeFragment.HomeFragment"
+        android:name="com.melvin.ongandroid.view.home.fragments.homeFragment.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" >
         <action
@@ -25,7 +25,7 @@
     </fragment>
     <fragment
         android:id="@+id/staffFragment"
-        android:name="com.melvin.ongandroid.view.fragments.staffFragment.StaffFragment"
+        android:name="com.melvin.ongandroid.view.home.fragments.staffFragment.StaffFragment"
         android:label="fragment_staff"
         tools:layout="@layout/fragment_staff" >
         <action
@@ -36,7 +36,7 @@
     </fragment>
     <fragment
         android:id="@+id/newsFragment"
-        android:name="com.melvin.ongandroid.view.fragments.newsFragment.NewsFragment"
+        android:name="com.melvin.ongandroid.view.home.fragments.newsFragment.NewsFragment"
         android:label="fragment_news"
         tools:layout="@layout/fragment_news" >
         <action
@@ -47,7 +47,7 @@
     </fragment>
     <fragment
         android:id="@+id/activitiesFragment"
-        android:name="com.melvin.ongandroid.view.fragments.activitiesFragment.ActivitiesFragment"
+        android:name="com.melvin.ongandroid.view.home.fragments.activitiesFragment.ActivitiesFragment"
         android:label="fragment_activities"
         tools:layout="@layout/fragment_activities">
         <action
@@ -58,7 +58,7 @@
     </fragment>
     <fragment
         android:id="@+id/contactFragment"
-        android:name="com.melvin.ongandroid.view.fragments.contactFragment.ContactFragment"
+        android:name="com.melvin.ongandroid.view.home.fragments.contactFragment.ContactFragment"
         android:label="fragment_contact"
         tools:layout="@layout/fragment_contact" >
         <action
